+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 ## [2.2] - Unreleased
 
+- New class `FieldOperator`. This class is an intermediate layer in the
+  inheritance hierarchy between `MonomialOperator` and
+  `CreationOperator`/`AnnihilationOperator`.
+
+- `GreensFunction` has been extended to compute anomalous GFs. This can be
+  achieved by passing two `AnnihilationOperator` objects to the constructor.
+  A unit test for this new functionality has been added.
+
 - New class `QuarticOperator`. In full analogy with `QuadraticOperator`, it is
   meant to be primarily used with `EnsembleAverage` to compute e.g. average
   double occupancy.